requirements.txt
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.24.3
opencv-python>=4.8.0.74
pydub>=0.25.1
imageio>=2.31.1
imageio-ffmpeg>=0.4.9
decorator>=4.4.2
proglog>=0.1.10
tqdm>=4.65.0
